feat: add error handling to prevent user from selecting invalid input value for numerical input ui

+++ ARC_Game_New/Assets/Scripts/SatisfactionAndBudget.cs
@@ -436,7 +436,7 @@ public class SatisfactionAndBudget : MonoBehaviour
         return currentSatisfaction;
     }
 
-    public float GetCurrentBudget()
+    public int GetCurrentBudget()
     {
         return currentBudget;
     }
